Resolved signed/unsigned warnings

+++ UnitTests/IntegrationTests_DoasFit.cpp
@@ -79,7 +79,7 @@ TEST_CASE("DoasFit - IntegrationTest with good scan - scan file 1", "[DoasFit][I
         // Change the settings to use HP500 filtering and make sure to filter the references
         // (but keep the unit in molec/cm2 as this makes it possible to have the same values in the assertions here as in the test above).
         so2FitWindow.fitType = FIT_TYPE::FIT_HP_DIV;
-        for (int refIdx = 0; refIdx < so2FitWindow.nRef; ++refIdx)
+        for (size_t refIdx = 0; refIdx < so2FitWindow.nRef; ++refIdx)
         {
             HighPassFilter(*so2FitWindow.ref[refIdx].m_data, CrossSectionUnit::cm2_molecule);
         }
@@ -119,7 +119,7 @@ TEST_CASE("DoasFit - IntegrationTest with good scan - scan file 1", "[DoasFit][I
         // Change the settings to use HP500 filtering and make sure to filter the references
         // (but keep the unit in molec/cm2 as this makes it possible to have the same values in the assertions here as in the test above).
         so2FitWindow.fitType = FIT_TYPE::FIT_HP_SUB;
-        for (int refIdx = 0; refIdx < so2FitWindow.nRef; ++refIdx)
+        for (size_t refIdx = 0; refIdx < so2FitWindow.nRef; ++refIdx)
         {
             HighPassFilter(*so2FitWindow.ref[refIdx].m_data, novac::CrossSectionUnit::cm2_molecule);
         }
